- fixed voting before 11 and ensured that all tests are run [INFO] Tests run: 91, Failures: 0, Errors: 0, Skipped: 0

+++ src/main/java/graduation/graduationProject/service/VoteService.java
@@ -31,6 +31,13 @@ public class VoteService {
 
     @Transactional
     public Vote create(int userId, int restaurant_id) {
+        if (get(userId) != null) {
+            if (isAfterEleven()) {
+                throw new ModificationRestrictionException();
+            } else {
+                crudVoteRepository.update(restaurant_id, userId, LocalDate.now());
+            }
+        }
         Vote vote = new Vote();
         vote.setUser(crudUserRepository.getOne(userId));
         vote.setRestaurant(crudRestaurantRepository.getOne(restaurant_id));
@@ -50,7 +57,14 @@ public class VoteService {
     }
 
     public boolean delete(int id, int userId) {
+        if (isAfterEleven()) {
+            throw new ModificationRestrictionException();
+        }
         checkNotFoundWithId(crudVoteRepository.delete(id, userId) != 0, id);
         return true;
+    }
+
+    private boolean isAfterEleven() {
+        return LocalTime.now().isAfter(LocalTime.of(11, 0));
     }
 }
